music: move TR2 music backends to TRX
