+++ src/libtrx/include/libtrx/game/music.h
@@ -3,3 +3,4 @@
 #include "./music/common.h"
 #include "./music/const.h"
 #include "./music/ids.h"
+#include "./music/types.h"
