+++ src/libtrx/include/libtrx/game/music/backend_files.h
@@ -1,5 +1,5 @@
 #pragma once
 
-#include "game/music/music_backend.h"
+#include "./types.h"
 
 MUSIC_BACKEND *Music_Backend_Files_Factory(const char *path);
