+++ src/libtrx/include/libtrx/game/music/types.h
