+++ src/tr2/game/music/music_main.c
@@ -1,14 +1,13 @@
 #include "game/music.h"
 
-#include "game/music/music_backend.h"
-#include "game/music/music_backend_cdaudio.h"
-#include "game/music/music_backend_files.h"
 #include "global/const.h"
 #include "global/types.h"
 #include "global/vars.h"
 
 #include <libtrx/config.h>
 #include <libtrx/engine/audio.h>
+#include <libtrx/game/music/backend_cdaudio.h>
+#include <libtrx/game/music/backend_files.h>
 #include <libtrx/log.h>
 
 static MUSIC_TRACK_ID m_TrackCurrent = MX_INACTIVE;
