+++ src/tr2/meson.build
@@ -128,8 +128,6 @@ sources = [
   'game/lara/state.c',
   'game/level.c',
   'game/los.c',
-  'game/music/music_backend_cdaudio.c',
-  'game/music/music_backend_files.c',
   'game/music/music_main.c',
   'game/objects/common.c',
   'game/objects/creatures/bandit_1.c',
